Run tests in separate processes.
