+++ src/lib.rs
@@ -63,9 +63,6 @@ pub mod glue;
 pub mod jsval;
 pub mod rust;
 
-#[cfg(test)]
-mod tests;
-
 pub use consts::*;
 
 use heapsize::HeapSizeOf;
